Speedup -> fixed, release build -> fixed, norm conv on terrain -> fixed (press L)

+++ TSBB11/src/camera.cpp
@@ -77,8 +77,8 @@ void Camera::change_look_at_pos(int xrel, int y, int width, int height)
         y = 1;
     }
     x += xrel;
-    float fi = ((float)x)/width*2*M_PI;
-    float theta = ((float)y)/height*M_PI;
+    float fi = ((float)x) / ((float)width) * 2.0f * M_PI;
+    float theta = ((float)y) / ((float)height) * M_PI;
 
     look_at_pos.x = -sin(theta)*sin(fi) + position.x;
     look_at_pos.y = cos(theta) + position.y;
